+++ test/getInternals.test.ts
@@ -1,5 +1,5 @@
 import { expect } from "@esm-bundle/chai";
-import { define, getInternals } from "../src/index.js";
+import { define, enhance, getInternals } from "../src/index.js";
 import { generateTagName } from "./helpers.js";
 const test = it;
 
@@ -27,6 +27,52 @@ describe("getInternals()", () => {
       const internals3 = getInternals(testEl);
       expect(internals1).to.equal(internals3);
       expect(() => testEl.attachInternals()).to.throw();
+    });
+
+    test("Using getInternals() in enhanced subclasses", () => {
+      @enhance()
+      class Base extends HTMLElement {}
+      class Test extends Base {
+        constructor() {
+          super();
+          getInternals(this);
+        }
+      }
+      window.customElements.define(generateTagName(), Test);
+      const testEl = new Test();
+    });
+
+    test("Using attachInternals() in enhanced subclasses", () => {
+      @enhance()
+      class Base extends HTMLElement {}
+      class Test extends Base {
+        constructor() {
+          super();
+          this.attachInternals();
+        }
+      }
+      window.customElements.define(generateTagName(), Test);
+      const testEl = new Test();
+    });
+
+    test("Using attachInternals() in a mixin in a decorator on a subclass", () => {
+      function whatever() {
+        return function (target: any) {
+          @enhance()
+          class Mixin extends target {}
+          return Mixin;
+        };
+      }
+      class Base extends HTMLElement {}
+      @whatever()
+      class Test extends Base {
+        constructor() {
+          super();
+          this.attachInternals(); // <- Error
+        }
+      }
+      window.customElements.define(generateTagName(), Test);
+      const testEl = new Test();
     });
   });
 });